keys: use Multikey for Activate, Single in Home
A previous refactor unified all keyfile generation to a single hook, useMultiKeyfileGenerator.

This turned out to cause issues as described in #895.

So, this commit resolves #895 by restoring single keyfile generation for Home view.

Tested on Ropsten, keyfiles generated in both Activation and Home views
