+++ src/views/UrbitOS/NetworkKeys.tsx
@@ -27,7 +27,7 @@ import {
 import useEthereumTransaction from 'lib/useEthereumTransaction';
 import { GAS_LIMITS, ONE_SECOND } from 'lib/constants';
 import { addHexPrefix } from 'lib/utils/address';
-import useMultikeyFileGenerator from 'lib/useMultikeyFileGenerator';
+import useKeyfileGenerator from 'lib/useKeyfileGenerator';
 
 import DownloadKeyfileButton from 'components/DownloadKeyfileButton';
 import InlineEthereumTransaction from 'components/InlineEthereumTransaction';
@@ -66,7 +66,7 @@ function useSetKeys(
     generating: keyfileGenerating,
     filename,
     bind: keyfileBind,
-  } = useMultikeyFileGenerator({ seed: manualNetworkSeed });
+  } = useKeyfileGenerator({ seed: manualNetworkSeed });
 
   const buildNetworkSeed = useCallback(
     async manualSeed => {
